added more recipe images, moved recipe objects to separate file, added a slug check for routing
